Add processing folder to .gitignore.
A sensible place is needed to store data which is not yet ready to be made into output i.e. data that needs to be preprocessed before its fed to one of the output functions contained in this repo. For this reason such a folder was made, and its name was added to the .gitignore file.